# Add checksum cache