--- a/superflore/generators/bitbake/run.py
+++ b/superflore/generators/bitbake/run.py
@@ -84,13 +84,13 @@ def main():
         try:
             md5_file = open('%s/md5_cache.pickle', 'rb')
             md5_cache = pickle.load(md5_file)
-            close(md5_file)
+            md5_file.close()
         except IOError:
             md5_cache = dict()
         try:
             sha256_file = open('%s/md5_cache.pickle', 'rb')
             sha256_cache = pickle.load(sha256_file)
-            close(sha256_file)
+            sha256_file.close()
         except IOError:
             sha256_cache = dict()
     with TempfileManager(args.output_repository_path) as _repo: