--- a/superflore/generators/bitbake/yocto_recipe.py
+++ b/superflore/generators/bitbake/yocto_recipe.py
@@ -54,30 +54,15 @@ class yoctoRecipe(object):
         self.archive_name = None
         self.license_md5 = None
         self.tar_dir = tar_dir
-        if md5_cache is not None:
-            if self.getArchiveName() not in md5_cache:
+        if self.getArchiveName() not in md5_cache and \
+           self.getArchiveName() not in sha256_cache:
                 self.downloadArchive()
                 md5_cache[self.getArchiveName()] = hashlib.md5(
                     open(self.getArchiveName(), 'rb').read()).hexdigest()
-                md5_file = open('%s/md5_cache.pickle' % tar_dir, 'wb')
-                pickle.dump(md5_cache, md5_file)
-                md5_file.close()
-            self.src_md5 = md5_cache[self.getArchiveName()]
-        else:
-            self.downloadArchive()
-            self.src_md5 = hashlib.md5(
-                open(self.getArchiveName(), 'rb').read()).hexdigest()
-        if sha256_cache is not None:
-            if self.getArchiveName() not in sha256_cache:
                 sha256_cache[self.getArchiveName()] = hashlib.sha256(
-                    open(self.getArchiveName(), 'rb').read()).hexdigest()
-                sha256_file = open('%s/sha256_cache.pickle' % tar_dir, 'wb')
-                pickle.dump(sha256_cache, sha256_file)
-                sha256_file.close()
-            self.src_sha256 = sha256_cache[self.getArchiveName()]
-        else:
-            self.src_sha256 = hashlib.sha256(
                 open(self.getArchiveName(), 'rb').read()).hexdigest()
+        self.src_sha256 = sha256_cache[self.getArchiveName()]
+        self.src_md5 = md5_cache[self.getArchiveName()]
 
     def getFolderName(self):
         return self.name.replace("-", "_") + "-" + str(self.version)