This is an experiment to speed up the generation of OE recipes.