@allenh1 2017-12-19
> Does computing the hashes really become the bottle neck for the builds?

I'm not sure. OE is significantly slower than the ebuild generation, so I'm trying to figure out the source of the slow down. This is really the only extra step that's not in the ebuild generator, so I'm trying to speed things up here. Could be this does nothing.

But this also gives us the ability to remove the tar archives (as only the checksums are needed).