@allenh1 2017-12-21
@tfoote Ok, this is the full changeset. Let me know what you think. I'm going to do a benchmark to see if this helps at all.

This holds the full path from the file, including the distro and version.