@allenh1 2017-12-21
Ok, here's the benchmark.

Without cache:
---------------------
 * real: 9 minutes and 3 seconds.
 * user: 17.7 seconds
 * sys: 3.304 seconds

With cache:
-----------------
 * real: 6 minutes and 37 seconds
 * user: 15.5 seconds
 * sys: 2.640 seconds


That's somewhat significant, especially in the real time. This also has the added advantage that we can remove the tars after the pickle cache generates.