@allenh1 2017-12-21
> Sorry this ended up with a lot of suggestions on a new approach with a HashCache class suggested to allow code reuse and increase readability/decrease branching.

@tfoote Upon implementing, this turned into the `CacheManager` class (which is just a bit more generic).